Class files renamed to proper conventions
